+++ src/Examples/TypedEncoding/Conversions.hs
@@ -26,7 +26,7 @@ import qualified Data.Text as T
 import qualified Data.ByteString as B
 
 -- $setup
--- >>> :set -XOverloadedStrings -XMultiParamTypeClasses -XDataKinds
+-- >>> :set -XOverloadedStrings -XMultiParamTypeClasses -XDataKinds -XTypeApplications
 --
 -- This module contains some ghci friendly values to play with.
 --
@@ -58,8 +58,9 @@ helloAsciiT = EnASCII.byteString2TextS helloAsciiB
 
 -- * Subsets
 
+
 helloUtf8B :: Enc '["r-UTF8"] () B.ByteString
-helloUtf8B = inject Proxy helloAsciiB
+helloUtf8B = inject helloAsciiB
 -- ^ To get UTF8 annotation, instead of doing this: 
 --
 -- >>> encodeFAll . toEncoding () $ "HeLlo world" :: Either EncodeEx (Enc '["r-UTF8"] () B.ByteString)
@@ -71,7 +72,7 @@ helloUtf8B = inject Proxy helloAsciiB
 --
 -- @inject@ method accepts proxy to specify superset to use.
 --
--- >>> displ $ inject (Proxy :: Proxy "r-UTF8") helloAsciiB
+-- >>> displ $ inject @ "r-UTF8" helloAsciiB
 -- "MkEnc '[r-UTF8] () (ByteString HeLlo world)"
 
 
@@ -153,11 +154,11 @@ lenientSomething = recreateAll . toEncoding () $ "abc==CB"
 -- * Flattening
 
 b64IsAscii :: Enc '["r-ASCII"] () B.ByteString
-b64IsAscii = flattenAs Proxy helloUtf8B64B
+b64IsAscii = flattenAs helloUtf8B64B
 -- ^ Base 64 encodes binary data as ASCII text. 
 -- thus, we should be able to treat "enc-B64" as "r-ASCII" losing some information.
 -- this is done using 'FlattenAs' type class
 --
--- >>> :t flattenAs (Proxy :: Proxy "r-ASCII") helloUtf8B64B
--- flattenAs (Proxy :: Proxy "r-ASCII") helloUtf8B64B
+-- >>> :t flattenAs @ "r-ASCII" helloUtf8B64B
+-- flattenAs @ "r-ASCII" helloUtf8B64B
 -- ... :: Enc '["r-ASCII"] () B.ByteString